feat(packages): allow for multiple rewards on one card
