--- packages/babylon-core-ui/src/widgets/sections/RewardSubsection/index.tsx
@@ -1,1 +1,0 @@
-export { RewardSubsection } from "./RewardSubsection";
