+++ packages/babylon-core-ui/src/widgets/sections/RewardsSubsection/index.tsx
@@ -1,0 +1,1 @@
+export { RewardsSubsection } from "./RewardsSubsection";
